Some adjustments in PDF.JS settings
So it'd work smoother on my old Nexus 4.

I've increased the data chunk size PDF.JS requests from the Android side
a little bit to make number of requests smaller.

Also, limited number of pixels on the canvas, it slightly decreases
sharpness of the text, but significanly improves performance on my old device.
I think it's a nice balance.