Fix the progress bar for the PDF documents
Ticket: https://github.com/hidroh/materialistic/issues/1012

The lifecycle of a PDF document in Materialistic is somewhat different
from a regular HTML page - in the HTML doc, we just open the doc. With
the PDF doc - we first need to download it, then initialize the
renderer, and only then - fetch the pieces we need to render the PDF
pages and actually render them.

Since the current progress bar is just showing loading of the HTML
document, it doesn't show any progress while downloading the PDF, which
is usually the longest part, especially on a slow mobile connection.
Because of that, users may make an assumption the page is broken or
unavailable, though it's just still downloading the PDf doc.

We could solve that by implementing a separate logic for the progress
bar while opening a PDF document. E.g.:

* We figure out this is a PDF doc - we set the progress bar to 10%.
* We finished downloading the PDF doc - we set the progress bar to 80%.
* We finished initializing PDF renderer, and started to fetch chunks of
  PDF doc via the bridge - we set the progress bar to 100%

We could go even further and implement downloading progress, so it'd
correctly progresses from 10 to 80 percent while downloading, but that
would require adding another network interceptor to `Call.Factory` and
way more plumbing than it's necessary for such a small feature.

So, I've done the following changes:

* Added `mIsPdf` field to `WebFragment`, which I use to suppress the
  HTML progress bar logic
* Added another callback to the bridge - onLoad, when it's called, I set
  the progress to 100% and hide the progress bar
* Added updating of the progress bar while processing the PDF doc
  according to the logic I described above
* Moved `PdfAndroidJavascriptBridgeFailureCallback` inside the bridge
  class to make the name shorter

Testing: Opened several HTML and PDF docs, observed the progress bar
behavior