Add proxy support
This commit adds support for proxying network requests over a SOCKS
or HTTP proxy and the corresponding preferences screen.

A DelegatingProxy (which is a java.net.Proxy) is registered on every
RestService. The DelegatingProxy delegates its proxying to a real
Proxy created from the proxy preferences.

This implementation is a bit ugly but the fact that RestServices are
only created once while proxy preferences can change at any time made
it hard to avoid global state and the DelegatingProxy class.

Enabling proxying breaks search.